refactor table into a multitude of components
