+++ src/table/body/table-expanded-row.component.ts
@@ -8,11 +8,7 @@ import {
 	// tslint:disable-next-line: component-selector
 	selector: "[ibmTableExpandedRow]",
 	template: `
-		<td
-			[ibmDataGridFocus]="isDataGrid"
-			[(columnIndex)]="columnIndex"
-			[attr.colspan]="row.length + 2"
-			(click)="setExpandIndex()">
+		<td [attr.colspan]="row.length + 2">
 			<ng-container *ngIf="!firstExpandedTemplateInRow(row)">
 				{{firstExpandedDataInRow(row)}}
 			</ng-container>
@@ -26,19 +22,11 @@ import {
 export class TableExpandedRow {
 	@Input() row: any[];
 
-	@Input() columnIndex = 0;
-
-	@Input() isDataGrid = false;
-
 	@Input() skeleton = false;
 
 	@HostBinding("class.bx--expandable-row") expandableRowClass = true;
 
 	@HostBinding("attr.data-child-row") @Input() expanded = false;
-
-	setExpandIndex() {
-		this.columnIndex = 0;
-	}
 
 	firstExpandedTemplateInRow(row) {
 		const found = row.find(d => d.expandedTemplate);
